# docs/openapi.yaml
openapi: "3.0.0"
info:
  title: WhatsApp API MultiDevice
  version: 5.0.0
  description: This API is used for sending whatsapp via API
servers:
  - url: http://localhost:3000
...
paths:
...
  /send/message:
    post:
      operationId: sendMessage
      tags:
        - send
      summary: Send Message
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                phone:
                  type: string
                  example: '6289685028129@s.whatsapp.net'
                  description: Phone number with country code
                message:
                  type: string
                  example: selamat malam
                  description: Message to send
                reply_message_id:
                  type: string
                  example: 3EB089B9D6ADD58153C561
                  description: Message ID that you want reply
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
  /send/image:
    post:
      operationId: sendImage
      tags:
        - send
      summary: Send Image
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                phone:
                  type: string
                  example: '6289685028129@s.whatsapp.net'
                  description: Phone number with country code
                caption:
                  type: string
                  example: selamat malam
                  description: Caption to send
                view_once:
                  type: boolean
                  example: false
                  description: View once
                image:
                  type: string
                  format: binary
                  description: Image to send
                compress:
                  type: boolean
                  example: false
                  description: Compress image
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
  /send/audio:
    post:
      operationId: sendAudio
      tags:
        - send
      summary: Send Audio
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                phone:
                  type: string
                  example: '6289685028129@s.whatsapp.net'
                  description: Phone number with country code
                audio:
                  type: string
                  format: binary
                  description: Audio to send
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
  /send/file:
    post:
      operationId: sendFile
      tags:
        - send
      summary: Send File
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                phone:
                  type: string
                  example: '6289685028129@s.whatsapp.net'
                  description: Phone number with country code
                caption:
                  type: string
                  example: selamat malam
                  description: Caption to send
                file:
                  type: string
                  format: binary
                  description: File to send
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
  /send/video:
    post:
      operationId: sendVideo
      tags:
        - send
      summary: Send Video
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                phone:
                  type: string
                  example: '6289685028129@s.whatsapp.net'
                  description: Phone number with country code
                caption:
                  type: string
                  example: ini contoh caption video
                  description: Caption to send
                view_once:
                  type: boolean
                  example: 'false'
                  description: View once
                video:
                  type: string
                  format: binary
                  description: Video to send
                compress:
                  type: boolean
                  example: 'false'
                  description: Compress video
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
  /send/contact:
    post:
      operationId: sendContact
      tags:
        - send
      summary: Send Contact
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                phone:
                  type: string
                  example: '6289685024051@s.whatsapp.net'
                  description: Phone number with country code
                contact_name:
                  type: string
                  example: Aldino Kemal
                  description: Contact name
                contact_phone:
                  type: string
                  example: '6289685024992'
                  description: Contact phone number
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
  /send/link:
    post:
      operationId: sendLink
      tags:
        - send
      summary: Send Link
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                phone:
                  type: string
                  example: '6289685024051@s.whatsapp.net'
                  description: Phone number with country code
                link:
                  type: string
                  example: "https://google.com"
                  description: Link to send
                caption:
                  type: string
                  example: 'Halo ini contoh caption'
                  description: Caption to send
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
  /send/location:
    post:
      operationId: sendLocation
      tags:
        - send
      summary: Send Location
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                phone:
                  type: string
                  example: '6289685024051@s.whatsapp.net'
                  description: Phone number with country code
                latitude:
                  type: string
                  example: "-7.797068"
                  description: Latitude coordinate
                longitude:
                  type: string
                  example: '110.370529'
                  description: Longitude coordinate
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
  /send/poll:
    post:
      operationId: sendPoll
      tags:
        - send
      summary: Send Poll / Vote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone:
                  type: string
                  description: The WhatsApp phone number to send the poll to, including the '@s.whatsapp.net' suffix.
                  example: '6289685024421@s.whatsapp.net'
                question:
                  type: string
                  description: The question for the poll.
                  example: 'Siapa Nama Avatar The Last Air Bender?'
                options:
                  type: array
                  description: The options for the poll.
                  items:
                    type: string
                  example: [ 'Zuko', 'Aang', 'Katara' ]
                max_answer:
                  type: integer
                  description: The maximum number of answers allowed for the poll.
                  example: 2
              required:
                - phone
                - question
                - options
                - max_answer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
  /send/presence:
    post:
      operationId: sendPresence
      tags:
        - send
      summary: Send presence status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                presence:
                  type: string
                  description: The presence status to send
                  enum: [available, unavailable]
                  example: 'available'
              required:
                - presence
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
...
components:
  schemas:
...
    SendResponse:
      type: object
      properties:
        code:
          type: string
          example: SUCCESS
        message:
          type: string
          example: Success
        results:
          type: object
          properties:
            message_id:
              type: string
              example: '3EB0B430B6F8F1D0E053AC120E0A9E5C'
            status:
              type: string
              example: '<feature> success ....'
...
    ErrorInternalServer:
      type: object
      properties:
        code:
          type: string
          example: INTERNAL_SERVER_ERROR
          description: 'SYSTEM_CODE_ERROR'
        message:
          type: string
          example: you are not loggin
          description: 'Detail error message'
        results:
          type: object
          example: null
          description: 'additional data'
    ErrorBadRequest:
      type: object
      properties:
        code:
          type: string
          example: 400
          description: 'HTTP Status Code'
        message:
          type: string
          example: field cannot be blank
          description: 'Detail error message'
        results:
          type: object
          example: null
          description: 'additional data'
...